backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
pillow==10.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
alembic==1.13.1
yomitoku==0.3.0
